Module Tiny_httpd

Tiny Http Server

This library implements a very simple, basic HTTP/1.1 server using blocking IOs and threads. Basic routing based on Scanf is provided for convenience, so that several handlers can be registered.

It is possible to use a thread pool, see create's argument new_thread.

The echo example (see src/examples/echo.ml) demonstrates some of the features by declaring a few endpoints, including one for uploading files:

module S = Tiny_httpd

let () =
  let server = S.create () in

  (* say hello *)
  S.add_route_handler ~meth:`GET server
    S.Route.(exact "hello" @/ string @/ return)
    (fun name _req -> S.Response.make_string (Ok ("hello " ^name ^"!\n")));

  (* echo request *)
  S.add_route_handler server
    S.Route.(exact "echo" @/ return)
    (fun req -> S.Response.make_string
        (Ok (Format.asprintf "echo:@ %a@." S.Request.pp req)));

  (* file upload *)
  S.add_route_handler ~meth:`PUT server
    S.Route.(exact "upload" @/ string_urlencoded @/ return)
    (fun path req ->
        try
          let oc = open_out @@ "/tmp/" ^ path in
          output_string oc req.S.Request.body;
          flush oc;
          S.Response.make_string (Ok "uploaded file")
        with e ->
          S.Response.fail ~code:500 "couldn't upload file: %s"
            (Printexc.to_string e)
      );

  (* run the server *)
  Printf.printf "listening on http://%s:%d\n%!" (S.addr server) (S.port server);
  match S.run server with
  | Ok () -> ()
  | Error e -> raise e

It is then possible to query it using curl:

$ dune exec src/examples/echo.exe &
listening on http://127.0.0.1:8080

# the path "hello/name" greets you.
$ curl -X GET http://localhost:8080/hello/quadrarotaphile
hello quadrarotaphile!

# the path "echo" just prints the request.
$ curl -X GET http://localhost:8080/echo --data "howdy y'all"
echo:
{meth=GET;
 headers=Host: localhost:8080
         User-Agent: curl/7.66.0
         Accept: */*
         Content-Length: 10
         Content-Type: application/x-www-form-urlencoded;
 path="/echo"; body="howdy y'all"}

Tiny buffer implementation

These buffers are used to avoid allocating too many byte arrays when processing streams and parsing requests.

module Buf = Tiny_httpd_buf

Generic byte streams

module Byte_stream = Tiny_httpd_stream

IO Abstraction

module IO = Tiny_httpd_io

Main Server Type

type buf = Tiny_httpd_buf.t

type byte_stream = Tiny_httpd_stream.t

module Meth = Tiny_httpd_server.Meth

module Headers = Tiny_httpd_server.Headers

module Request = Tiny_httpd_server.Request

module Response_code = Tiny_httpd_server.Response_code

module Response = Tiny_httpd_server.Response

module Route = Tiny_httpd_server.Route

module Middleware = Tiny_httpd_server.Middleware

type t = Tiny_httpd_server.t

val create : ?⁠masksigpipe:bool -> ?⁠max_connections:int -> ?⁠timeout:float -> ?⁠buf_size:int -> ?⁠get_time_s:(unit -> float) -> ?⁠new_thread:((unit -> unit) -> unit) -> ?⁠addr:string -> ?⁠port:int -> ?⁠sock:Unix.file_descr -> ?⁠middlewares:([ `Encoding | `Stage of int ] * Middleware.t) list -> unit -> t

module type IO_BACKEND = sig ... end

val create_from : ?⁠buf_size:int -> ?⁠middlewares:([ `Encoding | `Stage of int ] * Middleware.t) list -> backend:(module IO_BACKEND) -> unit -> t

val addr : t -> string

val is_ipv6 : t -> bool

val port : t -> int

val active_connections : t -> int

val add_decode_request_cb : t -> (unit Request.t -> (unit Request.t * (byte_stream -> byte_stream)) option) -> unit

val add_encode_response_cb : t -> (unit Request.t -> Response.t -> Response.t option) -> unit

val add_middleware : stage:[ `Encoding | `Stage of int ] -> t -> Middleware.t -> unit

val set_top_handler : t -> (byte_stream Request.t -> Response.t) -> unit

val add_route_handler : ?⁠accept:(unit Request.t -> (unit, Response_code.t * string) Stdlib.result) -> ?⁠middlewares:Middleware.t list -> ?⁠meth:Meth.t -> t -> ('a, string Request.t -> Response.t) Route.t -> 'a -> unit

val add_route_handler_stream : ?⁠accept:(unit Request.t -> (unit, Response_code.t * string) Stdlib.result) -> ?⁠middlewares:Middleware.t list -> ?⁠meth:Meth.t -> t -> ('a, byte_stream Request.t -> Response.t) Route.t -> 'a -> unit

module type SERVER_SENT_GENERATOR = sig ... end

type server_sent_generator = (module SERVER_SENT_GENERATOR)

val add_route_server_sent_handler : ?⁠accept:(unit Request.t -> (unit, Response_code.t * string) Stdlib.result) -> t -> ('a, string Request.t -> server_sent_generator -> unit) Route.t -> 'a -> unit

val running : t -> bool

val stop : t -> unit

val run : ?⁠after_init:(unit -> unit) -> t -> (unit, exn) Stdlib.result

val run_exn : ?⁠after_init:(unit -> unit) -> t -> unit

val _debug : ((('a, Stdlib.out_channel, unit, unit, unit, unit) Stdlib.format6 -> 'a) -> unit) -> unit

val _enable_debug : bool -> unit

Utils

module Util = Tiny_httpd_util

Resource pool

module Pool = Tiny_httpd_pool

Static directory serving

module Dir = Tiny_httpd_dir

module Html = Tiny_httpd_html

Alias to Tiny_httpd_html