Module Tiny_httpd.Request

HTTP request

A request sent by a client.

type 'body t = {

meth : Meth.t;

headers : Headers.t;

path : string;

body : 'body;

}

A request with method, path, headers, and a body.

The body is polymorphic because the request goes through several transformations. First it has no body, as only the request and headers are read; then it has a stream body; then the body might be entirely read as a string via read_body_full.

val pp : Format.formatter -> string t -> unit

Pretty print the request and its body

val pp_ : Format.formatter -> _ t -> unit

Pretty print the request without its body

val headers : _ t -> Headers.t

val get_header : ?⁠f:(string -> string) -> _ t -> string -> string option

val get_header_int : _ t -> string -> int option

val set_header : 'a t -> string -> string -> 'a t

val meth : _ t -> Meth.t

val path : _ t -> string

val body : 'b t -> 'b

val read_body_full : stream t -> string t

Read the whole body into a string. Potentially blocking.