Module Tiny_httpd.Response

type out_stream = bytes -> int -> int -> int

type t

val make_raw : ?⁠headers:Headers.t -> code:Response_code.t -> string -> t

val make_raw_chunked : ?⁠headers:Headers.t -> code:Response_code.t -> out_stream -> t

val make : ?⁠headers:Headers.t -> (string, Response_code.t * string) Pervasives.result -> t

val fail : ?⁠headers:Headers.t -> code:int -> ('a, unit, string, t) Pervasives.format4 -> 'a

Make the current request fail with the given code and message. Example: fail ~code:404 "oh noes, %s not found" "waldo"

val fail_raise : code:int -> ('a, unit, string, 'b) Pervasives.format4 -> 'a

Similar to fail but raises an exception that exits the current handler. This should not be used outside of a (path) handler. Example: fail_raise ~code:404 "oh noes, %s not found" "waldo"; never_executed()

val pp : Format.formatter -> t -> unit