Module Tiny_httpd_core.Buf

Simple buffer.

These buffers are used to avoid allocating too many byte arrays when processing streams and parsing requests.

type t

val size : t -> int

val clear : t -> unit

val create : ?size:int -> unit -> t

val of_bytes : bytes -> t

val contents : t -> string

val clear_and_zero : t -> unit

Clear the buffer and zero out its storage.

val bytes_slice : t -> bytes

Access underlying slice of bytes.

val contents_and_clear : t -> string

Get contents of the buffer and clear it.

val add_char : t -> char -> unit

Add a single char.

val add_bytes : t -> bytes -> int -> int -> unit

Append given bytes slice to the buffer.

val add_string : t -> string -> unit

Add string.

val add_buffer : t -> Stdlib.Buffer.t -> unit

Append bytes from buffer.