Module Tiny_httpd_html.Out

Output for HTML combinators.

This output type is used to produce a string reasonably efficiently from a tree of combinators.

NOTE: this is experimental and an unstable API.

Output for HTML combinators.

This output type is used to produce a string reasonably efficiently from a tree of combinators.

NOTE: this is experimental and an unstable API.

type t

val create_of_buffer : Stdlib.Buffer.t -> t

val create_of_out : Tiny_httpd_core.IO.Output.t -> t

val flush : t -> unit

val add_char : t -> char -> unit

val add_string : t -> string -> unit

val add_format_nl : t -> unit

val with_no_format_nl : t -> (unit -> 'a) -> 'a