Module Tiny_httpd_prometheus

Expose metrics over HTTP in the prometheus format.

This sub-library tiny_httpd.prometheus provides definitions for counters, gauges, and histogram, and endpoints to expose them for Prometheus to scrape them.

type tags = (string * string) list

module Registry : sig ... end

Registry for metrics.

val global : Registry.t

module Counter : sig ... end

Counters

module Gauge : sig ... end

Gauges

module Histogram : sig ... end

val http_middleware : Registry.t -> Tiny_httpd_core.Server.Middleware.t

Middleware to get basic metrics about HTTP requests

val add_route_to_server : Tiny_httpd_core.Server.t -> Registry.t -> unit

Add a "/metrics" route to the server

val instrument_server : Tiny_httpd_core.Server.t -> Registry.t -> unit

Add middleware and route

module GC_metrics : sig ... end