Module Tiny_httpd_prometheus.GC_metrics

type t

val create : Registry.t -> t

val update : t -> unit

val create_and_update_before_emit : Registry.t -> unit

create_and_update_before_emit reg creates new GC metrics, adds them to the registry, and uses Registry.on_will_emit to update the metrics every time the registry is polled.