Module Dir.Embedded_fs

An embedded file system, as a list of files with (relative) paths. This is useful in combination with the "tiny-httpd-mkfs" tool, which embeds the files it's given into a OCaml module.

type t

The pseudo-filesystem

val create : ?mtime:float -> unit -> t

val add_file : ?mtime:float -> t -> path:string -> string -> unit

Add file to the virtual file system.

raises Invalid_argument
if the path contains '..' or if it tries to make a directory out of an existing path that is a file.

val to_vfs : t -> (module VFS)