Module Tiny_httpd_ws

module IO = Tiny_httpd_io

type handler = Unix.sockaddr -> IO.Input.t -> IO.Output.t -> unit

Websocket handler

val upgrade : IO.Input.t -> IO.Output.t -> IO.Input.t * IO.Output.t

Upgrade a byte stream to the websocket framing protocol.

val add_route_handler : ?accept:(unit Tiny_httpd_server.Request.t -> (unit, int * string) result) -> ?accept_ws_protocol:(string -> bool) -> Tiny_httpd_server.t -> (Tiny_httpd_server.upgrade_handler, Tiny_httpd_server.upgrade_handler) Tiny_httpd_server.Route.t -> handler -> unit

Add a route handler for a websocket endpoint.

parameter accept_ws_protocol
decides whether this endpoint accepts the websocket protocol sent by the client. Default accepts everything.