Module Tiny_httpd_ws.With_lock

Synchronization primitive used to allow both the reader to reply to "ping", and the handler to send messages, without stepping on each other's toes.

type t = {

with_lock : 'a. (unit -> 'a) -> 'a;

}

A primitive to run the callback in a critical section where others cannot run at the same time.

The default is a mutex, but that works poorly with thread pools so it's possible to use a semaphore or a cooperative mutex instead.

type builder = unit -> t

val default_builder : builder

Lock using Mutex.