Package cmdliner 1.3.0

Cmdliner provides a simple and compositional mechanism to convert command line arguments to OCaml values and pass them to your functions.

The library automatically handles syntax errors, help messages and UNIX man page generation. It supports programs with single or multiple commands (like git) and respect most of the POSIX and GNU conventions.

Manuals

The following manuals are available.

The tutorial gets you through the steps to write your first command line interface with Cmdliner.

The Command line interface manual describes how command lines and environment variables are parsed by Cmdliner.

Tool man pages describes how Cmdliner generates man pages for your tools and how you can format them.

The examples page has a few annoted examples that show to express the command line interface of a few classic tools with Cmdliner

API

Cmdliner

Package info

authors

The cmdliner programmers

changes-files

CHANGES.md

depends

ocaml

homepage

https://erratique.ch/software/cmdliner

issues

https://github.com/dbuenzli/cmdliner/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/cmdliner/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/cmdliner.git

tags

cli

declarative

org:erratique

system

version

1.3.0