Package mtime 2.1.0

Mtime has platform independent support for monotonic wall-clock time. This time increases monotonically and is not subject to operating system calendar time adjustments. The library has types to represent nanosecond precision timestamps and time spans.

Mtime_clock provides acces to a monotonic system clock and its resolution (if available).

Library mtime

Mtime

Library mtime.clock.os

Mtime_clock

This library also works with JavaScript, see the platform support.

Package info

authors

The mtime programmers

changes-files

CHANGES.md

depends

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/mtime

issues

https://github.com/dbuenzli/mtime/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/mtime/doc/

readme-files

README.md

repo

git+https://erratique.ch/repos/mtime.git

tags

monotonic

org:erratique

system

time

version

2.1.0