Module Tiny_httpd

module Meth : sig ... end

module Headers : sig ... end

module Request : sig ... end

module Response_code : sig ... end

module Response : sig ... end

type t

val create : ?⁠masksigpipe:bool -> ?⁠fork:((unit -> unit) -> unit) -> ?⁠addr:string -> ?⁠port:int -> unit -> t

val addr : t -> string

val port : t -> int

val add_decode_request_cb : t -> (string Request.t -> string Request.t option) -> unit

Add a callback for every request. The callback can modify the request by returning Some r' where r' is the new request, or just perform side effects (logging?) and return None.

val add_encode_response_cb : t -> (string Request.t -> Response.t -> Response.t option) -> unit

Add a callback for every request/response pair. Similarly to add_encode_response_cb the callback can return a new response, for example to compress it.

val set_top_handler : t -> (string Request.t -> Response.t) -> unit

Setup a handler called by default. If not installed, unhandled paths will return a 404 not found.

val add_path_handler : ?⁠accept:(unit Request.t -> (unit, Response_code.t * string) Pervasives.result) -> ?⁠meth:Meth.t -> t -> ('a, Scanf.Scanning.in_channel, 'b, 'c -> string Request.t -> Response.t, 'a -> 'd, 'd) Pervasives.format6 -> 'c -> unit

add_path_handler server "/some/path/%s@/%d/" f calls f request "foo" 42 () when a request with path "some/path/foo/42/" is received. This uses Scanf's splitting, which has some gotchas (in particular, "%s" is eager, so it's generally necessary to delimit its scope with a "@/" delimiter. The "@" before a character indicates it's a separator.

parameter meth

if provided, only accept requests with the given method

parameter accept

should return true if the given request (before its body is read) should be accepted, false if it's to be rejected (e.g. because its content is too big, or for some permission error).

val stop : t -> unit

val run : t -> (unit, exn) Pervasives.result