Module Tiny_httpd_ws

Websockets for Tiny_httpd.

This sub-library (tiny_httpd.ws) exports a small implementation for a websocket server. It has no additional dependencies.

type handler = unit Tiny_httpd_core.Request.t -> Tiny_httpd_core.IO.Input.t -> Tiny_httpd_core.IO.Output.t -> unit

Websocket handler

val upgrade : Tiny_httpd_core.IO.Input.t -> Tiny_httpd_core.IO.Output.t -> Tiny_httpd_core.IO.Input.t * Tiny_httpd_core.IO.Output.t

Upgrade a byte stream to the websocket framing protocol.

exception Close_connection

Exception that can be raised from IOs inside the handler, when the connection is closed from underneath.

val add_route_handler : ?accept:(unit Tiny_httpd_core.Request.t -> (unit, int * string) result) -> ?accept_ws_protocol:(string -> bool) -> ?middlewares:Tiny_httpd_core.Server.Head_middleware.t list -> Tiny_httpd_core.Server.t -> (Tiny_httpd_core.Server.upgrade_handler, Tiny_httpd_core.Server.upgrade_handler) Tiny_httpd_core.Route.t -> handler -> unit

Add a route handler for a websocket endpoint.

parameter accept_ws_protocol
decides whether this endpoint accepts the websocket protocol sent by the client. Default accepts everything.